gcp(phase-d): review remediation — poll terminality + module hardening
Two-agent adversarial review of the Phase D delta merged under live-test
pressure (backend: SHIP, no HIGH/MED; modules: SHIP-WITH-FIXES).

- MED: poll stranded cancelled/stillborn Cloud Run executions in
  'running' forever — a cancelled run sets completionTime with
  succeeded=0/failed=0 (increments cancelledCount, not failedCount),
  which fell into the 'running' else-branch, so a wait-until-done loop
  (exit on done==true) never exited. Terminality now keys on
  completionTime: any completed execution is terminal; completed with no
  successes → 'failed'. Surface cancelledCount.
- LOW: gcp-run-logs-fetch clamped MAX_ENTRIES with (raw as u32) BEFORE
  clamp — a value >u32::MAX wrapped into range instead of clamping to
  the cap (lint-check-21 wraparound). Clamp in u64 first.
- LOW: gcp-run-job-execute reported started:true with an empty
  execution id when the run LRO carried no metadata name — now fails
  loud (poll/logs can't chain a blank id).
- LOW (backend, defense-in-depth): impersonation mint expiry fallback
  now uses the refresh margin, not the full lifetime, so an
  unconfirmed-expiry token (Google omitting expireTime) is never served
  stale from cache — it re-mints next dispatch.

Verified correct by review, no change needed: Logging-filter injection
(strict name validation before interpolation), no token/{:?} leak,
x-goog-user-project header injection blocked by HeaderValue validation,
cache insert-only-on-success bounds it against random-SA DoS,
refresh/revoke lockstep. check-catalog 69/69; talos-google-cloud green.

+++ module-templates/gcp-run-logs-fetch/template.rs
@@ -147,7 +147,10 @@ fn resolve_max_entries(config: &serde_json::Value) -> u32 {
         .get("MAX_ENTRIES")
         .and_then(|v| v.as_u64().or_else(|| v.as_str().and_then(|s| s.parse::<u64>().ok())))
         .unwrap_or(DEFAULT_MAX_ENTRIES as u64);
-    (raw as u32).clamp(MIN_ENTRIES, MAX_ENTRIES_CAP)
+    // Clamp in u64 BEFORE narrowing — casting first would wrap a value above
+    // u32::MAX down into range (e.g. 2^32+5 → 5) instead of clamping to the
+    // cap (integer-wraparound class, lint check 21).
+    raw.clamp(MIN_ENTRIES as u64, MAX_ENTRIES_CAP as u64) as u32
 }
 
 #[talos_module(world = "http-node")]
